requirements.txt
django
dice
inflect
pyyaml